update message_listener to async
